[CL-3258] Update seatInfo tests for i1
